Laget en components folder under blankett for å samle kode som tilhører blankett
