+++ src/server/blankett/components/blankett/AvslåVedtak.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { avslagÅrsakTilTekst, type IAvslåVedtak } from '../../../typer/dokumentApi';
+import { avslagÅrsakTilTekst, type IAvslåVedtak } from '../../../../typer/dokumentApi';
 
 export const AvslåVedtak: React.FC<IAvslåVedtak> = ({ avslåÅrsak, avslåBegrunnelse }) => {
   return (
