+++ src/server/blankett/components/blankett/Dokument.tsx
@@ -11,13 +11,13 @@ import SagtOppEllerRedusertGrunnlag from './SagtOppEllerRedusertGrunnlag';
 import { Vedtak } from './Vedtak';
 import AlderPåBarnGrunnlag from './AlderPåBarnGrunnlag';
 import { TidligereHistorikk } from './TidligereHistorikk';
-import type { IDokumentData, IVurdering, IVilkårGrunnlag } from '../../../typer/dokumentApi';
+import type { IDokumentData, IVurdering, IVilkårGrunnlag } from '../../../../typer/dokumentApi';
 import {
   VilkårGruppe,
   Vilkår,
   EBehandlingÅrsak,
   vilkårTypeTilTekst,
-} from '../../../typer/dokumentApi';
+} from '../../../../typer/dokumentApi';
 
 interface DokumentProps {
   dokumentData: IDokumentData;
