+++ src/server/hentDokumentHtmlBlankett.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 
-import Dokument from './components/blankett/Dokument';
+import Dokument from './blankett/components/blankett/Dokument';
 import { renderToStaticMarkup } from 'react-dom/server';
 import css from './utils/css';
-import Header from './components/blankett/Header';
+import Header from './blankett/components/blankett/Header';
 import { datoFormat } from './utils/util';
-import { Behandling } from './components/blankett/Behandling';
-import { ÅrsakRevurdering } from './components/blankett/ÅrsakRevurdering';
+import { Behandling } from './blankett/components/blankett/Behandling';
+import { ÅrsakRevurdering } from './blankett/components/blankett/ÅrsakRevurdering';
 import type { IDokumentData } from '../typer/dokumentApi';
 import { stønadstypeTilTekst } from '../typer/dokumentApi';
